+++ sxt/multiexp/bucket_method/accumulation_kernel.cc
@@ -14,4 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "sxt/multiexp/bucket_method/accumulate_kernel.h"
+#include "sxt/multiexp/bucket_method/accumulation_kernel.h"
